STOCK-20 python 코드 정리 및 trello연동
